requirements.txt
altgraph==0.17
asgiref==3.3.1
cart==1.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.4
django-db-signals==0.1.1
django-heroku==0.3.1
django-imagekit==4.0.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
future==0.18.2
geojson==2.5.0
gunicorn==20.0.4
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
pefile==2019.4.18
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.10.1
pygame==2.0.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==2.0.1
pyowm==2.10.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.25.0
requests-oauthlib==1.3.0
signals==0.0.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
urllib3==1.26.2
wf-datetime-conversion==0.0.1
whitenoise==5.2.0

django-modeltranslation~=0.17.2
pip~=20.2.3
setuptools~=49.2.1